added-which

+++ eoas.py
@@ -25,21 +25,28 @@ def add_zczc(text):
     return re.sub(r"(\n\n|\A\n*)(\d+)", r"\1ZCZC \2", text)
 
 
-def fetch(date=None, proxy=None, scan=None):
+def fetch(which=None, date=None, proxy=None, scan=None):
     """Fetches all EOAS files from the syn folder, prefixes ZCZCnnn to each message,
     and writes the combined files as YYYY-MM-DDgts.txt.
+    :param which: what subfolder to get; defaults to "syn"
+        can be "all" to get syn, metar, and aircraft
     :param date: the date to get; defaults to "today"
     :param proxy: optional host:port to use as proxy
     :param scan: will rescan every <scan> hours for new info
     """
+    which = which or "syn"
     while scan:
-        fetch(date, proxy, scan=False)
+        fetch(which, date, proxy, scan=False)
         print(f"sleeping {scan} hours")
         time.sleep(scan*3600)
+    if which == "all":
+        for which in ("syn", "metar", "aircraft"):
+            fetch(which, date, proxy)
+        return
     maxhrs, maxt, maxfile = 0, None, None
     date = date_of(date) or datetime.date.today()
-    print("fetching", date)
-    fn = date.strftime("%Y-%m-%d-EOAS-GTS.txt")
+    print("fetching", date, which)
+    fn = date.strftime("%Y-%m-%d-EOAS-{which}-GTS.txt")
     if os.path.exists(fn):
         os.remove(fn)
     session = requests.Session()
@@ -48,8 +55,8 @@ def fetch(date=None, proxy=None, scan=None):
             proxy = f"http://{proxy}"
         session.proxies = dict(http=proxy, https=proxy)
     # todo is a list of all the URLs we want to fetch
-    todo = ["http://rawdata.eoas.fsu.edu/syn/"]
-    wanted = re.compile(f"{date.strftime('%Y%m%d')}/$|\.syn$")
+    todo = [f"http://rawdata.eoas.fsu.edu/{which}/"]
+    wanted = re.compile(f"{date.strftime('%Y%m%d')}/$|\.[a-z]+$")
     while todo:
         url = todo.pop()
         try:
@@ -60,8 +67,8 @@ def fetch(date=None, proxy=None, scan=None):
         if not page.ok:
             print(f"got status {page.status_code} for {url}; skipping")
             continue
-        # we got a syn file, save it
-        if url.endswith(".syn"):
+        # if it's a file, save it
+        if re.search("\.[a-z]+$", url):
             cs = hashlib.md5(page.content).digest()
             old = md5s.get(url)
             if not old:
@@ -80,14 +87,14 @@ def fetch(date=None, proxy=None, scan=None):
                 continue
             # if u is wanted, put it on the todo list
             if wanted.search(u):
-                if u.endswith('.syn'):
+                if re.search("\.[a-z]+$", u):
                     t1 = datehour_of(e.text[:10])
                     t2 = datehour_of(e.xpath("./ancestor::td[1]/following-sibling::td[1]")[0].text)
                     hrs = (t2-t1).total_seconds()/3600
                     if hrs > maxhrs:
                         maxhrs, maxt, maxfile = hrs, t2, e.text
                 todo.append(u)
-    if scan is None:
+    if maxfile and scan is None:
         print(f'oldest update is {maxfile} updated {maxt}, {maxhrs:0.1f} hours later')
